Phase3/dashboard/requirements.txt
streamlit>=1.29.0
kafka-python>=2.0.2
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
